Change homepage, add header/footer to all pages
