+++ src/app/layout.tsx
@@ -1,9 +1,10 @@
 import type { Metadata } from "next";
-import { Inter, Montserrat } from "next/font/google";
+import { Montserrat } from "next/font/google";
 import "./globals.css";
 import { ClerkProvider } from "@clerk/nextjs";
+import Header from "@/components/header";
+import Footer from "@/components/footer";
 
-const inter = Inter({ subsets: ["latin"] });
 const montserrat = Montserrat({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
@@ -19,7 +20,11 @@ export default function RootLayout({
     return (
         <ClerkProvider>
             <html lang="en">
-                <body className={montserrat.className}>{children}</body>
+                <body className={montserrat.className}>
+                    <Header />
+                    {children}
+                    <Footer />
+                </body>
             </html>
         </ClerkProvider>
     );
